FEAT: userSeq로 사용자 정보 찾기 API

+++ server/rhythm4cuts/src/main/java/com/b109/rhythm4cuts/model/service/UserServiceImpl.java
@@ -391,7 +391,7 @@ public class UserServiceImpl implements UserService {
         return null;
     }
 
-    public String findNicknameById(int userSeq) {
+    public String findNicknameById(Long userSeq) {
         return userRepository.findByUserSeq(userSeq).getNickname();
     }
